// packages/mpl-core-kit-lib/src/generated/types/baseDataSection.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { type Codec, combineCodec, type Decoder, type Encoder, getStructDecoder, getStructEncoder } from '@solana/kit'

import {
  type BaseLinkedDataKey,
  type BaseLinkedDataKeyArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
  getBaseLinkedDataKeyDecoder,
  getBaseLinkedDataKeyEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
} from './index.ts'

export type BaseDataSection = {
  parentKey: BaseLinkedDataKey
  schema: ExternalPluginAdapterSchema
}

export type BaseDataSectionArgs = {
  parentKey: BaseLinkedDataKeyArgs
  schema: ExternalPluginAdapterSchemaArgs
}

export function getBaseDataSectionCodec(): Codec<BaseDataSectionArgs, BaseDataSection> {
  return combineCodec(getBaseDataSectionEncoder(), getBaseDataSectionDecoder())
}

export function getBaseDataSectionDecoder(): Decoder<BaseDataSection> {
  return getStructDecoder([
    ['parentKey', getBaseLinkedDataKeyDecoder()],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ])
}

export function getBaseDataSectionEncoder(): Encoder<BaseDataSectionArgs> {
  return getStructEncoder([
    ['parentKey', getBaseLinkedDataKeyEncoder()],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ])
}